requirements.txt
torch
torchvision
transformers
opencv-python
pillow
numpu
